html/semantics/forms/the-fieldset-element/fieldset-setcustomvalidity.html

<!DOCTYPE HTML>
<title>fieldset setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<fieldset id='test'></fieldset>

<script>

test(() => {
  let fieldset = document.getElementById("test");
  assert_false(fieldset.validity.customError);
  fieldset.setCustomValidity("custom error");
  assert_true(fieldset.validity.customError);
}, "fieldset setCustomValidity is correct")

</script>
